diary

+++ src/main/java/Hogwarts/controller/ScenicSpotController.java
@@ -18,7 +18,7 @@ public class ScenicSpotController {
     @PostMapping("")
     public ScenicSpot create(@RequestParam("name") String name, @RequestParam("rate") int rate,
                              @RequestParam("image") String image, @RequestParam("audio") String audio,
-                             @RequestParam("lat") float lat, @RequestParam("lng") float lng,
+                             @RequestParam("lat") double lat, @RequestParam("lng") double lng,
                              @RequestParam("count") int count, @RequestParam("heat") int heat,
                              @RequestParam("profile") String profile) {
         return scenicSpotService.create(name,rate,image,audio,lat,lng,count,heat,profile);
