+++ src/main/java/Hogwarts/service/ScenicSpotService.java
@@ -3,8 +3,8 @@ package Hogwarts.service;
 import Hogwarts.entity.ScenicSpot;
 
 public interface ScenicSpotService {
-    public ScenicSpot create(String name, int rate, String image, String audio, float lat,
-                             float lng, int count, int heat, String profile);
+    public ScenicSpot create(String name, double rate, String image, String audio, double lat,
+                             double lng, int count, int heat, String profile);
 
     //void updatehot(int sid);
 }
